When saving fetched entries in the db, check for each one if it's already deleted and if so ignore it instead of trying to save it.
Before this commit, this was taken care of by ActiveRecord validations. This meant that when fetching a feed with already deleted entries, the app worked as expected (entries were not saved again in the db) but the log was filled up with errors because of the validation errors being raised.

Instead, it's better to explicitly check if each entry is already deleted (a similar validation was already checking if the entry already existed in the db). This way the entry is simply ignored with a debug-level log message, instead of raising an error and logging an error-level message.