The main app menu in smartphone-sized viewports is aligned to the right.
It made no sense to have it aligned to the left when the button that makes the menu visible is at the right of the navbar. It forced the user to move his finger too much.

Also, minor changes: fixed-width fontawesome icons for better menu alignment, set submenus background to a dark color so that options in the submenu are easily distinguishable from options in the main menu.