Switched the navbar buttons in smartphones (main menu button is at the left, show/hide sidebar button at the right). Align text in main menu to the left.
Aligning text in smartphone menu to the right did not give a good result, the menu looked like a mess. Left-alignment is better. To keep the user from having to move his finger across the screen when opening the menu, the button has been moved to the left so that the user only has to move the fingertip vertically to access menu options.

This enables us to remove enpty icon placeholders (<i class="fa fa-fw"></i>) that were just used for menu alignment when the menu text was right-aligned.

Also minor changes to the menu separator in the folders dropdown: the separator only appears if the user has created at least one folder (it makes no sense to show the separator if there's nothing to separate) and in smartphone browsers make the separator darker (so that it's visible against the dark grey background of submenus).