The job that performs scheduled feed updates now validates that the feed is marked as available.
Feed updates should be unscheduled when a job is marked as unavailable. However if for some reason this fails and an unavailable feed still has scheduled updates, the next time ScheduledUpdateFeedJob runs for that feed, this will be detected and instead of trying to fetch the feed its updates will be unscheduled.

This fixes a rare bug in which a permanently unavailable feed still had scheduled updates, which were always failing and filling the logs with error messages.