Fixed folder and arrow icons displaying incorrectly after clicking on a sidebar folder.
It seems in recent angularjs versions passing a hash to an ng-class directive with multiple CSS classes in a hash key is not supported; in my case the first class ("fa") was being ignored, which resulted in fontawesome icons not being displayed correctly.

The fix is to make the "fa" class a permanent part of the "class" attribute value, and make a single class optional: "fa-folder-o", "fa-folder-open-o", etc.