"Share on twitter" button is now a link that opens in a new page.
This normally makes no difference because the twitter widget js opens a new popup when the button is clicked, regardless of the target attribute. However if the user is blocking the twitter widget library (e.g. he uses the Ghostery firefox plugin) the link becomes a regular link to a twitter page; in this case it's better that the link opens a new tab.