Pass an option hash so that the tooltip for entry buttons is attached to the page body.
This removes a bug in some browsers (at least Chrome) in which the buttons moved a bit to the left on mouseover.

See discussions about this issue:

https://github.com/twbs/bootstrap/issues/6811
https://github.com/twbs/bootstrap/pull/6378

The bootstrap docs, by the way, describe the "container" option for the "tooltip" function, but they don't really explain what it does nor that it fixes this issue.