Merged social share buttons in a single "share" dropdown menu in the entry toolbar.
This will make it easier to add other social networks in the future, it looks less cluttered, and it gives less problems with responsivity (the previous row of buttons didn't fit in a single row in smartphones).

The price to pay is that the user has to make one additional click to share entries. However I think it's a worthy tradeoff, because I guess most users will share far less entries than they read.