Finishing touches for the invitation feature.
Fixed error in the URL for the accept invitation form.

Removed the "destroy invitation" route. Users won't be able to destroy invitations; rather, unaccepted invitations will be destroyed automatically after a certain amount of time (this is not yet implemented).

Only the create invitation action is handled by our customized FriendsInvitationsController. The rest of actions are handled by the default devise_invitable controller.