When user invites an already invited user, the same invitation email is sent again.
This increments the invitations_count attribute of the inviter by 1. Not yet implemented: check that the inviter has enough invitations left before resending the invitation.

This is different from the default devise_invitable behavior, which is to generate a new invitation token when resending. I don't like this because it means that the invited user has several invitation emails, only one of them has an "accept" link that actually works, and it's not obvious which is the correct one.