Rewrite has_invitations_left? method in FriendInvitationsController to adapt to our needs.
We want the user always to be able to send invitations if he's an admin, or if he has no invitations limit set (invitation_limit set to null). In any other case we want the user to be able to send invitations only if he has invitations_count < invitation_limit.

Before resending an invitation, the invitation limit is also checked. Resending an invitation increments by 1 the invitations_count attribute, which keeps users from spamming an email address with invitations.

Reorganized unit test for FriendInvitationsController.