When first sending an invitation, persist the unencrypted invitation token in the "unencrypted_invitation_token" attribute of the invited User instance.
This is necessary to resend invitations without generating a new invitation token and invalidating the new one. This would be undesirable for us, because it would invalidate older invitations emails without it being obvious for the inviter.

Unfortunately devise_invitable has no native way to persist the unencrypted token in the db. Luckily it implements an instance method that can be used to retrieve the raw token just after sending the invitation, which we can use to roll out our own persistence logic.