Increment by 1 the inviter's invitation count when first sending an invitation to an email address.
I was trusting that devise_invitable already was doing this, but it isn't. It's probably because of the config in the initializer, but I'm not gonna mess with it because I intend to manage invitations limit a bit differently from what devise_invitable assumes by default. It's simple enough to do it in our customized FriendInvitationsController.