Allow invited users to sign up normally instead of accepting the invitation.
When a user who has been invited signs up instead of accepting the invitation, the user record associated with his email that was created  when the invitation was sent is destroyed, and then the signup is processed normally.

Before this commit users who had been invited could not sign up, because an "email is taken" error was raised.