Finished work in acceptance tests for the send invitation feature.
Also DRYed up tests: added support methods to sign up a new user account and accept an invitation to join Feedbunch during acceptance testing. These methods assume the successful path, which means that they can be used only in methods that have a successful signup or accepted invitation as part of the testing, not in methods that test some of the error paths in either signup or invitation accepting. A lot DRYer anyway.