Renamed Devise::FriendInvitationsController -> FeedbunchAuth::InvitationsController and Devise::ProfilesController -> FeedbunchAuth::RegistrationsController
This way the controller_name method of those controllers returns "invitations" and "registrations", exactly the same as the original Devise controllers they extend. This is desirable because Devise uses the controller_name method to compose the i18n keys for its flash messages; this way we can override the Devise controllers but keep the same i18n files (with their original keys).

This fixes "missing translation" messages that were appearing in the flash since implementing a custom Registrations controller.