Clear the ActionMailer deliveries queue before each test.
I'm hoping this will fix intermittent failures in the invitations acceptance test. I'm guessing those failures are caused by sent emails remaining in the queue when a test finishes, which would make tests that check that no email is sent fail afterwards. This would explain why the failures are intermittent, because they would be dependant on test ordering, which is random.