Wrapped checkboxes in edit registration view with a responsive grid div.
This way checkboxes have some left margin for viewport widths under the xs breakpoint (specifically, the checkboxes are offset 1 column to the right, so that 1 of the 12 boostrap grid columns is blank an to the left of the checkboxes). Above the xs breakpoint the form looks exactly the same as before, which is all right.