Use the same route ("/invitation") for all actions related to the invitation feature.
Using a different route ("/accept_invitation") for the GET action caused problems, because if the user caused an error when posting the form (e.g. leaving both passwords fields blank), the response redirected the user to "/invitation". This was a problem because this route was not accessible through GET, which meant that refreshing the page caused an error.