Added job to clean up invitations.
Currently this job (which is not yet scheduled anywhere) destroys User records (along with their associations) which correspond to invitations sent a long time ago which have not yet been accepted.

The interval after which unaccepted invitations will be destroyed can be configured in the discard_unaccepted_invitations_after variable in application.rb. The default value is 1 month.

This means that after 1 month invitations sent by email will no longer be valid.

In the future this job will probably reset the daily invitations limit per user.

Not yet implemented: statically schedule this job.