Fixed loading feeds when clicking on the top "all subscriptions" link to read all subscribed feeds.
When the user clicks on this link, all feeds are retrieved via ajax in a paginated fashion, and unread counts are updated in the view. Previously there was a bug: if the ajax call for the first page of results was returned a 404 nothing was done, and the view was unchanged. The expected behavior is that all unread counts are set to zero (because getting a404 in the first page means that there are no feeds with unread entries at all).

Bug fixed, regression test implemented.