When clicking on a folder during acceptance testing, the test waits for feeds and folders to finish loading before clicking on the folder.
This is intended to avoid an error in which capybara clicked on the "all subscriptions" link (to read all subscribed feeds) while feeds and folder hadn't finished loading, which made the test fail because no entries etc were loaded. The fix is to wait for feeds/folders to be loaded before clicking on any folder.

As a side note, this behavior (allowing the user to click on "all subscriptions" before feeds have finished loading) is a bug, because it gives unexpected results (most of the time clicking on the link does nothing, instead of loading all entries). But it will be fixed in a separate issue and commit.