Updated rest-client gem 1.6.7 -> 1.7.1
Since rest-client 1.7.0 when a connection times out a RestClient::RequestTimeout is raised instead of an Errno::ETIMEDOUT. Support for this new exception has been added; the change is not strictly necessary because it extends RestClient::Exception, which was already rescued in all cases, but it's good to be explicit in the code. The Errno::ETIMEDOUT is still rescued; maybe it's no longer raised, but it's better to be cautious in case there's a combination of circumstances that could still raise it.