Clicking on the "All subscriptions" link before any feeds have been loaded does nothing.
The exception is if feed pagination has finished without loading any feeds (i.e. the user is not subscribed to any feeds with unread entries). In this case the link can be clicked once pagination has finished.

The reason is that if we allow the user to click on the link before any feeds have been loaded, the entries list will be left blank which is confusing for the user. It's better to keep him from clicking on the link while it's not yet functional.