Removed bootstrap panel classes from entries list. Opening and closing entries is done with a ng-show directive, instead of adding and removing the "in" class (which is associated to bootstrap collapsible panels).
The reason is that we use very little of the bootstrap panels features. It's more convenient to use the ng-show directive to open/close entries instead of ng-class; this way we have support for animations from angularjs with minimal work.

The entries list looks a bit different (thinner borders, square instead of rounded corners in each entry) but actually I think it looks better now.