Scroll automatically when opening an entry using jQuery.animate() instead of angularjs $anchorScroll.
The scrolling is not instantaneous but animated over a few milliseconds, which looks much nicer. This cannot be easily achieved with angularjs alone.

A small delay is introduced before scrolling to give time enough for angularjs to close any other entries. This is necessary because if we try to scroll immediately, it's possible that angularjs has not finished closing other entries and the entry top would not be in its final position; this would mean that the page would scroll to a wrong position.

The invisible anchor element that was used for scrolling with $anchorScroll has been removed, it is no longer necessary with jquery. This also means that the browser address bar is not changed when opening an entry (previously a fragment was added to the URL).

Entry open-close animations have been temporarily disabled to implement this change in scrolling.