When clicking on sidebar links, load entries after loading feeds.
This way in most cases the client will detect that no entries have been received for a given feed, and its unread count will be set to zero even if a non-zero count is returned by the server.

A wrong count may still be displayed if the Ajax call to retrieve feeds is slower than the one to retrieve entries, but this should be rare.