When clicking on a folder, after all entries have been retrieved, any feeds in the folder which did not return any entries have their unread counts set to zero.
This is intended to correct (or at least hide from the user) those cases in which the unread count saved in the FeedSubscription instance is incorrect. E.g. the user marks all entries in a feed as read, but because the server is under load the job that actually updates the unread count takes some minutes to run.

Before this commit, if during those minutes the user clicked on the folder containing the feed, the entries list would contain no entries from that feed but the unread count in the sidebar for the feed would be non-zero. This commit fixes this: if no entries are received for a feed when clicking on its folder, the feed's unread count is set to zero in the client.

This kind of errors would be corrected when the UpdateFeedUnreadCountJob runs, or at the latest during the next scheduled update of the feed. This commit does nothing about actually fixing the data, it just makes the client hide the error when it is detected.