When the user clicks on the "accept" button of the cookies warning message, a cookie is set to remember this fact.
If this cookie is set to "true", the cookies warning message won't be displayed.

Currently it is a session cookie. Pending: study if the $cookies angularjs service can be used to set persistant cookies. Also consider if changes to Path, HttpOnly and Security attributes of the cookie would be desirable.