Set accept-cookies cookie to expire in the very far future (year 9.999).
It seems that Chrome ignores this date and instead sets the cookie to expire on 'Tue, 19 Jan 2038 03:14:07 GMT', which is the maximum date since Epoch expressed as an amount of seconds stored in a signed 32-bit integer. This shouldn't pose no problems.

Also set the Path explicitly to '/'. This should not be a problem, but it's better to be explicit.