Every time CleanupInvitationsJob runs (it currently is scheduled daily), the daily invitations limit for all users is set to the currently configured one.
The limit is configured in application.rb, in the daily_invitations_limit config value.

This enables us to change the configured limit and, after a restart, the next time the job runs this limit will be updated for all users. This enables changing the limit without manually making changes in the database.

This limit is ignored for admin users, who can send an unlimited number of invitations.