Every time CleanupInvitationsJob runs, the daily invitations count for each user is reset to zero.
This completes the daily invitation limit feature. Previous commits had set limits to prevent a user from sending an invitation if all his invites had been used. This commits adds a daily reset to zero of the invitations count, which means that the invitations limit for each user is actually a daily limit.

Logic has been added to ensure that the count is not reset before one day has passed, regardless of how many times the job runs (because of deployments etc). This is necessary because otherwise users could have the count reset multiple times a day, which would result in the limit not being enforced. Also users that have their current invitations count at zero (they sent no invitations during the last day) are ignored, for performance reasons.