Do not stub FeedClient.fetch in all ScheduledUpdateFeedJob tests.
It is not really necessary to stub it in all tests; RestClient.get is stubbed in spec_helper.rb, which ensures that no HTTP calls are made during testing. Those tests in scheduled_update_feed_job_spec.rb that need to stub FeedClient.fetch do it explicitly as part of the test.

Also, I will need to invoke the actual implementation of FeedClient.fetch to test a new feature: if a scheduled feed update fails, the job will attempt to perform autodiscovery, in case the feed URL has changed but the website URL remains the same (which in my experience happens quite often).