If FeedClient.fetch is invoked with the perform_autodiscovery argument set to true, feed.url will be passed to RestClient.get instead of feed.fetch_url. Also no HTTP caching headers will be sent.
This is intended for the case in which a scheduled update of the feed fails. In this case we want to attempt autodiscovery using feed.url; attempting it on feed.fetch_url is useless because that URL is failing. feed.url should lead to the webpage, which is our best bet to get the updated fetch_url if it has changed.

Also when we are performing autodiscovery we don't want to receive a "304 not modified" response, we want the last version of the actual HTML document to be able to do autodiscovery. So now the method does not send any HTTP caching headers if it's been told to perform autodiscovery.