When a scheduled feed update fails, attempt to perform autodiscovery on the website URL.
This is done only if the update failure raises an error that could mean that the URL is unavailable (timeout, host not found, etc). Other kind of errors that could arise e.g. from bugs in the app or problems in the server do not result in an autodiscovery attempt.

If a different fetch_url is discovered from the website, the fetch_url attribute of the feed is updated with the new value and a fetch is attempted from this url, which could result in new entries and in an updated feed title.

This is intended for the case in which a feed owner decides to change its URL, e.g. because he migrates from an in-house solution to feedburner. In these cases usually the website URL does not change, and often the site owner does not even give a warning that the old feed URL is no longer valid. Hopefully he will remember to update the autodiscovery information in his website with the new feed URL.

The old behavior was: the scheduled feed updates started failing, and after 1 week the feed would be marked as permenently unavailable. The feedbunch admin had this week to become aware of the problem and fix it manually by going to the website, copying the new feed URL and updating the feed with it using ActiveAdmin. It was a less than ideal solution that required too much manual intervention.

With the new behavior, feedbunch will try to autodetect the new feed url when this happens. As long as the feed owner updates the autodiscovery information in his website in the week following the URL change, feedbunch will update the feed data and mark the feed as "not failing" automatically, without manual intervention.

Of course if the site owner does not update the autodiscovery data or does not even support autodiscovery, this will fail and manual intervention will still be required. I'd like to send people who do this on a one-way trip to a special circle of hell without internet access, but alas, I have to suffer living in an imperfect world.