After the entry opening animation has finished, set its height to 'auto'.
This way, if images in the entry are lazy-loaded after the animation has finished, the entry height will automatically grow to adapt to the newly loaded entry.

Previous to this entry height was fixed, which meant that if images finished loading after the animation finished, not all content would fit in the allocated viewport space.