The 'entry_open' CSS class identifies whether an entry is currently open or not.
This class is inially added to all entries if the user has selected the "open all entries by default" config option.

Otherwise it is added when an entry is opened, and removed when the entry is closed (because the user clicked to close it, or because he opened another entry which triggered an automatic closing of all other open entries).