Smaller top offset when autoscrolling the page after opening an entry.
The offset was a bit too big, two entries were visible above the currently open one (at viewport widths above the md- cutoff). This is too much screen space not dedicated to the currently open entry.

This change means that for viewport widths under the md- cutoff, only part of the previous entry is visible above the currently open one. I think it's a worthy tradeoff.