When opening an entry, the scrolling to show as much of its content as possible is performed with velocity.js.
Using Velocity.js "scroll" animation is supposed to have better performance and use hardware acceleration in mobile devices. See: http://julian.com/research/velocity/#scroll

The call to velocity("scroll") is chained to the call that animates opening the entry. It is supposed to give better performance this way, and it eliminates any timing problems (the scroll animation is guaranteed to start immediately when the open animation finishes).

This removes the need for a separate scrollSvc service. The ng-ScrollSvc.js.coffee file has been removed along with any references to scrollSvc.