Fix velocity.js arguments syntax.
I was trying to use the single-object argument syntax, but there was an error that prevented the complete function from being invoked when the open entry animation finished. As a result the entry_open CSS class was never added to open entries, which caused problems (tests failing, entries with images overflowing their height).

I've gone to the two-objects arguments syntax (one object for CSS parameters, a second object for Velocity.js options). It seems to be the better supported syntax (I also had to use it in the end to make the scroll animation work).