When opening a folder, the check to see if any action is necessary is based in comparing folder IDs, not folder objects.
The reason is that folders may be reloaded from the server (e.g. when a feed is moved to a new folder) which creates a new collection of folder objects in memory. In this circumstances we must compare IDs to see if two objects represent the same folder; comparing the objects themselves could give a false negative if one folder object was loaded on page load, and the other one was loaded from the server at a later time.