Handle the situation in which the feed server resets the connection during a scheduled feed update.
In this case an Errno::ECONNRESET error will be raised. This error was not being rescued, which meant that the error bubbled to Resque and the update failure logic (mark the feed as failing etc) was not being executed.

The fix is to add this error to the rescue clauses.