findSvc angularjs service now can search a folder in a list of folders, instead of searching in the list of folders stored in the root scope.
If no list is passed to the service, the default behavior is still to search in the root scope, so this shouldn't break anything.

This will enable us to search in the list of folders returned by the server. This will be necessary to change the behavior when loading feeds: instead of just replacing the list of folders in the root scope with the folders returned by the server (which will replace all folder objects with new ones), only new folders will be inserted in the scope (and folders in the scope not present in the list returned by the server will be removed). This way folders that have not changed will not have the object that represents them replaced by a new identical one.