When loading folders from the server, insert only new folders, and remove any folders not returned by the server.
The old behavior was simply to replace the full list of folders with the one returned by the server. This meant that even for folders that had not changed, the object representing them in the scope was replaced by a new one. This caused complications keeping track of which folder was open in the sidebar at a given time.

With the new behavior, folders that haven't changed will not change in the angularjs model layer. This makes more sense, even if the price to pay is a slightly more complicated logic.