Specify z-index for entries and sidebar columns only in smartphone-sized screens.
These z-indexes are necessary for the sidebar to be able to completely hide the entries column, when the sidebar is visible.

For other screen sizes this caused problems (the feed title tooltip was partially hidden by the navbar). Not specifying a z-index for bigger screens works; in smartphones the problem with tooltips does not happen because tooltips are not displayed in smartphones.