Added Fastclick JS library.
This library should eliminate the 300ms delay in touchscreens. This delay is intended to make sure the user is not double-clicking, but in apps like this which use no double-clicking it just makes the app feel unresponsive.

Fastclick is supposed to attach no listeners if it's not needed (e.g. in desktop browsers), so no loss of performance is expected.

For more about Fastclick:

https://github.com/ftlabs/fastclick