Improve the client-side code that keeps unread counts up to date in the browser.
When all entries from a feed or folder have been loaded, unread entries actually received for each feed are counted and the client-side unread count is updated. This ensures that the unread is (more or less) correct when pagination finishes. Unfortunately it may be wrong until the last page of entries has been loaded; this is unavoidable, the client simply does not know how many unread entries are in not-yet-received pages.

No action is done to notify the server of any error in unread counts. Scheduled jobs are supposed to run and fix this information as needed.

Previously a limited functionality was implemented: when loading a folder's entries, feeds which had returned no entries got their unread count set to zero. This commit implements a superset of that: now it works both when reding a feed and when reading a folder, and also the unread count is corrected for all feeds, not just for those that have no unread entries.

Also a bug has been fixed: the old implementation counted entries without regard to their read/unread state; this meant that if "show read" had been selected, the unread count could very well be wrong because it included read entries.